requirements.txt
torch
matplotlib
numpy
pandas
pixeltable-yolox